+++ server/src/modules/quest/quest.controller.ts
@@ -211,23 +211,4 @@ export class QuestController {
   ): Promise<void> {
     await this.questService.deleteSubQuest(user.id, id);
   }
-
-  @Patch('side/:id')
-  @UseGuards(JwtAuthGuard)
-  @ApiOperation({ summary: '사이드 퀘스트 상태 수정' })
-  @ApiBearerAuth('accessToken')
-  @ApiQuery({ name: 'id', type: Number, description: '사이드 퀘스트 ID' })
-  @ApiBody({ type: UpdateQuestStatusRequest })
-  @ApiNoContentResponse({ description: 'success' })
-  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
-  @ApiNotFoundResponse({ description: 'fail - Quests not found' })
-  @ApiInternalServerErrorResponse({ description: 'Internal Server Error' })
-  @HttpCode(HttpStatus.NO_CONTENT)
-  async updateSideQuestStatus(
-    @CurrentUser() user: AccessTokenPayload,
-    @Param('id', ParseIntPipe) id: number,
-    @Body() request: UpdateQuestStatusRequest
-  ) {
-    await this.questService.updateSideQuestStatus(user.id, id, request);
-  }
 }
