+++ server/src/modules/quest/quest.module.ts
@@ -2,19 +2,11 @@ import { Module } from '@nestjs/common';
 import { QuestService } from './services/quest.service';
 import { QuestController } from './quest.controller';
 import { QuestRepositoryModule } from '@entities/quest/quest-repository.module';
-import { SideQuestRepositoryModule } from '@entities/side-quest/side-quest-repository.module';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Quest } from '@entities/quest/quest.entity';
-import { SideQuest } from '@entities/side-quest/side-quest.entity';
-import { SideQuestService } from './services/side-quest.service';
+import { SideQuestModule } from '@modules/side-quest/side-quest.module';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([Quest, SideQuest]),
-    QuestRepositoryModule,
-    SideQuestRepositoryModule,
-  ],
+  imports: [QuestRepositoryModule, SideQuestModule],
+  providers: [QuestService],
   controllers: [QuestController],
-  providers: [QuestService, SideQuestService],
 })
 export class QuestModule {}
